sides fixed

+++ JMM/mesh2D.h
@@ -74,11 +74,11 @@ void printGeneralInfoMesh(mesh2S *mesh2);
 void printEverythingInMesh(mesh2S *mesh2);
 
 void twoTrianglesFromEdge(mesh2S *mesh2, size_t index0, size_t index1,
-			  size_t possibleTriangles[2], size_t possibleThirdVertices[2] );
+			  int possibleTriangles[2], size_t possibleThirdVertices[2] );
 
 double minEtaFromTwoPoints(mesh2S *mesh2, size_t index0, size_t index1);
 
-size_t faceBetween3Points(mesh2S *mesh2, size_t index0, size_t index1, size_t index2);
+int faceBetween3Points(mesh2S *mesh2, size_t index0, size_t index1, size_t index2);
 
 
 void triangleFan_initFromIndices(triangleFanS *triFan, mesh2S *mesh2, size_t nRegions,
